refactor: move some next pages helpers to utilities package
Move getNextPages, sanitizePagePaths and getNextPagesRoot to utilities/src/utils/common
